Improved Gizmo handling. Now modules can add gizmos.
