+++ modules/godot/editor_plugins/components_gizmo_3d.h
@@ -17,10 +17,15 @@ struct RelativeHandle {
 class Components3DGizmoPlugin : public EditorNode3DGizmoPlugin {
 	GDCLASS(Components3DGizmoPlugin, EditorNode3DGizmoPlugin);
 
+	static Components3DGizmoPlugin *singleton;
+
 	LocalVector<Ref<ComponentGizmo>> gizmos;
 
 public:
+	static Components3DGizmoPlugin *get_singleton() { return singleton; }
+
 	Components3DGizmoPlugin();
+	~Components3DGizmoPlugin();
 
 	/// Add a component gizmo.
 	void add_component_gizmo(Ref<ComponentGizmo> p_gizmo);
@@ -61,43 +66,4 @@ public:
 	Ref<StandardMaterial3D> get_material(const String &p_name, const Ref<EditorNode3DGizmo> &p_gizmo = Ref<EditorNode3DGizmo>()) {
 		return owner->get_material(p_name, p_gizmo);
 	}
-};
-
-class TransformComponentGizmo : public ComponentGizmo {
-	StringName transform_component_name = "TransformComponent";
-
-public:
-	virtual void init() override;
-	virtual void redraw(EditorNode3DGizmo *p_gizmo) override;
-	virtual int get_handle_count() const override;
-	virtual String get_handle_name(const EditorNode3DGizmo *p_gizmo, int p_idx) const override;
-	virtual Variant get_handle_value(EditorNode3DGizmo *p_gizmo, int p_idx) const override;
-	virtual void set_handle(EditorNode3DGizmo *p_gizmo, int p_idx, Camera3D *p_camera, const Point2 &p_point) override;
-	virtual void commit_handle(EditorNode3DGizmo *p_gizmo, int p_idx, const Variant &p_restore, bool p_cancel = false) override;
-};
-
-class MeshComponentGizmo : public ComponentGizmo {
-	StringName mesh_component_name = "MeshComponent";
-	StringName transform_component_name = "TransformComponent";
-	StringName transform_name = "transform";
-
-	struct EditorMeshData : public ComponentGizmoData {
-		RID base;
-		RID instance;
-
-		EditorMeshData();
-		~EditorMeshData();
-		void set_mesh(Ref<Mesh> p_mesh);
-
-		virtual void on_position_update(const Transform &p_new_transform) override;
-	};
-
-public:
-	virtual void init() override;
-	virtual void redraw(EditorNode3DGizmo *p_gizmo) override;
-	virtual int get_handle_count() const override;
-	virtual String get_handle_name(const EditorNode3DGizmo *p_gizmo, int p_idx) const override;
-	virtual Variant get_handle_value(EditorNode3DGizmo *p_gizmo, int p_idx) const override;
-	virtual void set_handle(EditorNode3DGizmo *p_gizmo, int p_idx, Camera3D *p_camera, const Point2 &p_point) override;
-	virtual void commit_handle(EditorNode3DGizmo *p_gizmo, int p_idx, const Variant &p_restore, bool p_cancel = false) override;
 };
