+++ register_types.cpp
@@ -7,9 +7,14 @@
 #include "editor/plugins/node_3d_editor_plugin.h"
 #include "iterators/dynamic_query.h"
 #include "modules/ecs_modules_register.h"
+#include "modules/godot/editor_plugins/components_gizmo_3d.h"
 #include "systems/dynamic_system.h"
 
+Ref<Components3DGizmoPlugin> component_gizmo;
+
 void register_godex_types() {
+	component_gizmo.instance();
+
 	godex::DynamicSystemInfo::for_each_name = StringName("_for_each");
 
 	ClassDB::register_class<ECS>();
@@ -30,6 +35,8 @@ void register_godex_types() {
 }
 
 void unregister_godex_types() {
+	component_gizmo = Ref<Components3DGizmoPlugin>();
+
 	ecs_unregister_modules();
 
 	// Clear dynamic system static memory.
